vis-short/requirements.txt
streamlit
altair
pandas
pymssql
python-dotenv